Fixing Consrate, QtyRen Daan, PriceDaan

+++ app/Http/Controllers/ConsRateController.php
@@ -13,17 +13,13 @@ use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Bus;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Validator;
 use Maatwebsite\Excel\Facades\Excel;
 
 class ConsRateController extends Controller
 {
     public function index(Request $request, ConsRateDataTable $consRateDataTable)
     {
-//        $data = Asumsi_Umum::query()
-//            ->select('version_asumsi.id', 'version_asumsi.version', DB::raw("date_part( 'year' :: TEXT, asumsi_umum.month_year )||'-'|| lpad(date_part( 'month' :: TEXT, asumsi_umum.month_year)::TEXT, 2, '0') as month_year"), DB::raw("date_part( 'year' :: TEXT, asumsi_umum.saldo_awal )||'-'|| lpad(date_part( 'month' :: TEXT, asumsi_umum.asumsi_umum.saldo_awal)::TEXT, 2, '0') as saldo_awal"))
-//            ->leftJoin('version_asumsi', 'version_asumsi.id', '=', 'asumsi_umum.version_id')->get();
-
-//        dd($data);
         if ($request->data == 'index') {
             return $consRateDataTable->render('pages.buku_besar.consrate.index');
         }
@@ -34,7 +30,8 @@ class ConsRateController extends Controller
     public function create(Request $request)
     {
         try {
-            $request->validate([
+
+            $validator = Validator::make($request->all(), [
                 "id_plant" => 'required',
                 "version" => 'required',
                 "id_asumsi" => 'required',
@@ -42,7 +39,10 @@ class ConsRateController extends Controller
                 "material" => 'required',
                 "consrate" => 'required',
                 "is_active" => 'required',
-            ]);
+            ], validatorMsg());
+
+            if ($validator->fails())
+                return $this->makeValidMsg($validator);
 
             $data_asumsi = Asumsi_Umum::where('id', $request->id_asumsi)
                 ->first();
@@ -84,7 +84,7 @@ class ConsRateController extends Controller
     public function update(Request $request)
     {
         try {
-            $request->validate([
+            $validator = Validator::make($request->all(), [
                 "id_plant" => 'required',
                 "version" => 'required',
                 "id_asumsi" => 'required',
@@ -92,8 +92,11 @@ class ConsRateController extends Controller
                 "material" => 'required',
                 "consrate" => 'required',
                 "is_active" => 'required',
-            ]);
-            //            dd($request);
+            ], validatorMsg());
+
+            if ($validator->fails())
+                return $this->makeValidMsg($validator);
+
 
             if (strpos($request->id_asumsi, '-') == true) {
                 $data_asumsi = Asumsi_Umum::where([
@@ -148,12 +151,15 @@ class ConsRateController extends Controller
     public function import(Request $request)
     {
         try {
-            DB::transaction(function () use ($request) {
-                $request->validate([
-                    'file' => 'required',
-                    'version' => 'required'
-                ]);
+            $validator = Validator::make($request->all(), [
+                "file" => 'required',
+                "version" => 'required',
+            ], validatorMsg());
 
+            if ($validator->fails())
+                return $this->makeValidMsg($validator);
+
+            DB::transaction(function () use ($request) {
                 ConsRate::where('version_id', $request->version)->delete();
 
                 $file = $request->file('file')->store('import');
